r192: allocate mi_idx_t::seq and name from kalloc
This changes aims to reduce the malloc overhead. In the previous version, the
peak memory dramatically increases for the second batch of sequences to be
indexed. This is due to a memory leak. The peak memory does not increase with
the third batch and so on. I speculate the observation is caused by libc malloc
reserving memory for scalability.

It is difficult to address this issue in the current code. This commit only
slightly reduces peak memory. I will think more about this at a later time.